+++ src/climb.rs
@@ -4,21 +4,21 @@ use crate::fit_reader;
 use crate::gem_finder;
 use crate::math;
 
-fn _get_climb(
+fn get_climb(
     section: &dtypes::WindowSection,
     altitudes: &dtypes::Altitudes,
     times: &dtypes::Times,
 ) -> f64 {
-    let gained_altitude_in_section: f64 = _get_gained_altitude_in_section(
+    let gained_altitude_in_section: f64 = get_gained_altitude_in_section(
         &altitudes.values,
         section.start as usize,
         section.end as usize,
     );
     let duration = times.values[section.end as usize] - times.values[section.start as usize];
-    math::_climb_equation(&gained_altitude_in_section, &(duration / 60.))
+    math::climb_equation(&gained_altitude_in_section, &(duration / 60.))
 }
 
-fn _get_gained_altitude_in_section(altitudes: &Vec<f64>, start: usize, end: usize) -> f64 {
+fn get_gained_altitude_in_section(altitudes: &Vec<f64>, start: usize, end: usize) -> f64 {
     let mut section = altitudes[start..end].to_vec();
     // drop NAN values
     section.retain(|&i| i.is_normal());
@@ -38,7 +38,7 @@ pub fn update_sections_max_climb(
 ) {
     window_sec.distance = input_data.distances.values[window_sec.end as usize]
         - input_data.distances.values[window_sec.start as usize];
-    window_sec.climb = _get_climb(&window_sec, &input_data.altitudes, &input_data.times);
+    window_sec.climb = get_climb(&window_sec, &input_data.altitudes, &input_data.times);
     // update fastest_sec only in case the current distance is equal to the desired distance +- 1% and velocity is larger
     if gem_finder::distance_in_bounds(
         window_sec.distance,
@@ -55,10 +55,13 @@ pub fn update_sections_max_climb(
 pub fn specific_data_check(
     input_data: &gem_finder::InputData,
 ) -> Result<(), errors::InputDataError> {
-    if input_data.altitudes.values.len() < 2 {
-        return Err(errors::InputDataError::TooFewDataPoints);
-    } else if input_data.coordinates.len() != input_data.altitudes.values.len() {
+    if input_data.coordinates.len() != input_data.altitudes.values.len() {
         return Err(errors::InputDataError::InconsistentLength);
+    }
+    let mut altitudes_normal = input_data.altitudes.values.clone();
+    altitudes_normal.retain(|&i| i.is_normal());
+    if altitudes_normal.len() < 2 {
+        return Err(errors::InputDataError::TooFewDataPoints);
     } else {
         return Ok(());
     }
@@ -80,10 +83,11 @@ pub fn find_best_climb_section(
     ) {
         Err(e) => Err(e),
         Ok(mut finder) => {
-            finder._compute_vector_of_distances();
-            match finder._check_if_total_distance_suffice() {
-                Ok(_) => match finder._search_section(update_sections_max_climb) {
-                    Ok(result) => return Ok(result),
+            math::fill_nans(&mut finder.coordinates);
+            finder.compute_vector_of_distances();
+            match finder.check_if_total_distance_suffice() {
+                Ok(_) => match specific_data_check(&finder) {
+                    Ok(_) => return finder.search_section(update_sections_max_climb),
                     Err(e) => return Err(e),
                 },
                 Err(e) => return Err(e),
@@ -115,7 +119,7 @@ mod test_climb {
     #[test]
     fn test_get_gained_altitude_in_section_all_values() {
         let altitudes = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-        let result = _get_gained_altitude_in_section(&altitudes, 0, altitudes.len());
+        let result = get_gained_altitude_in_section(&altitudes, 0, altitudes.len());
         let expected = 4.0;
         assert_eq!(expected, result);
     }
@@ -123,7 +127,7 @@ mod test_climb {
     #[test]
     fn test_get_gained_altitude_in_section_slice_only() {
         let altitudes = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-        let result = _get_gained_altitude_in_section(&altitudes, 1, altitudes.len() - 1);
+        let result = get_gained_altitude_in_section(&altitudes, 1, altitudes.len() - 1);
         let expected = 2.0;
         assert_eq!(expected, result);
     }
@@ -131,7 +135,7 @@ mod test_climb {
     #[test]
     fn test_get_gained_altitude_in_section_also_going_downhill() {
         let altitudes = vec![1.0, 2.0, 3.0, 4.0, 5.0, 4.0, 3.0];
-        let result = _get_gained_altitude_in_section(&altitudes, 0, altitudes.len());
+        let result = get_gained_altitude_in_section(&altitudes, 0, altitudes.len());
         // expect 4.0, since only climbing uphill counts, negative values (going downhill) will be dropped
         let expected = 4.0;
         assert_eq!(expected, result);
@@ -140,7 +144,7 @@ mod test_climb {
     #[test]
     fn test_get_gained_altitude_in_section_including_nan() {
         let altitudes = vec![1.0, 2.0, 3.0, f64::NAN, 4.0];
-        let result = _get_gained_altitude_in_section(&altitudes, 0, altitudes.len());
+        let result = get_gained_altitude_in_section(&altitudes, 0, altitudes.len());
         // expect 3.0 since the nan value will be dropped
         let expected = 3.0;
         assert_eq!(expected, result);
@@ -151,7 +155,7 @@ mod test_climb {
     #[test]
     fn test_find_best_climb_section_in_fit() {
         let result = find_best_climb_section_in_fit(1_000., FIT_FILE, Some(0.01)).unwrap();
-        assert_eq!(result.start, 344);
+        assert_eq!(result.start, 346);
         assert_eq!(result.end, 586);
         assert_eq!(result.target_value.round(), 6.0);
     }
@@ -160,7 +164,46 @@ mod test_climb {
     fn test_find_best_climb_section_in_fit_larger_section() {
         let result = find_best_climb_section_in_fit(3_000., FIT_FILE, Some(0.01)).unwrap();
         assert_eq!(result.start, 63);
-        assert_eq!(result.end, 708);
+        assert_eq!(result.end, 705);
         assert_eq!(result.target_value.round(), 4.0);
+    }
+}
+
+#[cfg(test)]
+mod test_checks {
+    use super::*;
+
+    #[test]
+    fn test_inconsistent_length_of_altitude_data() {
+        let coordinates = vec![(1., 1.), (2., 2.)];
+        let times = vec![1., 2.];
+        let altitudes = vec![123., 124., 125.];
+        assert_eq!(
+            find_best_climb_section(
+                1_000.,
+                coordinates,
+                times,
+                altitudes,
+                Some(gem_finder::DEFAULT_TOLERANCE)
+            ),
+            Err(errors::InputDataError::InconsistentLength)
+        )
+    }
+
+    #[test]
+    fn test_nans_in_altitude() {
+        let coordinates = vec![(1., 1.), (2., 2.)];
+        let times = vec![1., 2.];
+        let altitudes = vec![123., f64::NAN];
+        assert_eq!(
+            find_best_climb_section(
+                1_000.,
+                coordinates,
+                times,
+                altitudes,
+                Some(gem_finder::DEFAULT_TOLERANCE)
+            ),
+            Err(errors::InputDataError::TooFewDataPoints)
+        )
     }
 }